♻️(front) make the resourceList reducer generic
Now that the resourceById reducer and the resourceList saga are
generic, all we have to do is make the resourceList reducer generic
so we can fetch lists of any resource.
